Make list of copyrights deterministic
When printing out the list of copyrights in a file, they are not in any
specific order.  So when the tool runs on the same repo multiple times,
the output is not in any specific order, making future diffs to the BOM
a bit hard to understand why they are made.

To solve this, always output the list of copyrights in sorted order.