Refactor _verify_paths_supported() in two methods
I've done this refactor to try to make the code more clear. Now the two
responsibilities of the method are split and the conditions to use them
are more explicit in run(). The cost is having to loop twice over the
paths, but I've prefered clarity over a possibly despicable loss of
performance.