Ignore go.sum files
go.sum is the machine-maintained half of Go Modules, and is a list of
checksums that should be checked into source control [1]. Any time the
module is changed or a dependency of the module is updated, this file is
clobbered. It is arguably a build artifact that is checked into source
control for reproducability, and thus can't be put in .gitignore.

[1]: https://blog.golang.org/using-go-modules