default_template: follow Kernel style
The Kernel suggests to put the SPDX license identifier as high up as
possible, reflect that in the default_template.

[...]
The SPDX license identifier in kernel files shall be added at the first
possible line in a file which can contain a comment. For the majority or
files this is the first line, except for scripts which require the
‘#!PATH_TO_INTERPRETER’ in the first line. For those scripts the SPDX
identifier goes into the second line.
[...]

Signed-off-by: Paul Spooren <mail@aparcar.org>