Comment style for uncommentable files
Previously, only binary files were handled as uncommentable [0]. For
other files the "--explicit-license" flag has to be set explicitly.
However, there are also text files which are uncommentable.

This commit / pull request introduces support for a new pseudo comment
style for uncommentable text files.

One use case is, for example, Golang's go.sum file. Hashes of the
referenced modules are listed there, but without any special header [1].
Thus it is not possible to add an SPDX header. However, this file cannot
be omitted. Whether the file is to be ignored license-wise was addressed
in a previous pull request [2], but not clarified.

Another prominent use case might be JSON files, which also cannot
contain comments or headers. This was also discussed before [3].

With the changes introduced here, this file is recognized as
uncommentable and a .license file is created.

[0] https://reuse.software/tutorial/#binary-and-uncommentable-files
[1] https://blog.golang.org/using-go-modules
[2] https://github.com/fsfe/reuse-tool/pull/234
[3] https://github.com/fsfe/reuse-tool/issues/189