feat: support comment style for .bashrc files
This commit ensures that the .bashrc filename is supported for the addheader
command.

The .bashrc files are used to configure the bash terminal. Typically they are
included from other bash configuration files. It is similar to a bash
script like the already supported .bash extension. In this case however the
hashbang will typically not be there.

Signed-off-by: Nico Rikken <nico.rikken@alliander.com>