feat: ignore case when matching files
Changes the default behavior of the matching of filenames and extensions to be
case-insensitive. His is done by casting the COMMENT_STYLE_MAPS's and the files
to be matched to lowercase. I've chosen this runtime casting approach over
redefining the maps in lowercase to have the filenames and extensions in their
more common form, which can be capitalized like 'Dockerfile' or 'MANIFEST.in'.
If we need to get back on the capitalization behavior, we would still have the
definitions in a proper format.

Fixes #355

Signed-off-by: Nico Rikken <nico.rikken@alliander.com>