Recommend installation via pipx
Managing python, pip and virtual environments can be too much work sometimes especially for a project that is not itself written in python. pipx is a convenient shortcut, but people unfamiliar with the python ecosystem might not even know it exists. To mitigate this I have added an example of running your tool via pipx to your readme.