Add Nix-related files
This adds compatibility with Nix, the purely functional package manager.
This gives us a declarative development environment with `reuse` in
editable mode and a fully reproducible build of `reuse` and its entire
dependency tree across multiple architectures. As a cherry on top, it
can also build a fully reproducible docker image. If you are curious
about Nix, the package manager, check out the manual:

https://nixos.org/manual/nix/stable/

For more information about Flakes, this blog post is helpful:

https://www.tweag.io/blog/2020-05-25-flakes/