feat: ansible.cfg comment style support
Support the INI format syntax for `ansible.cfg` files. This file is used
to configure the Ansible software and can reside in multiple locations.
The INI format uses the hash sign (#) for comments that occupy an entire
line. The PythonCommentStyle is chosen to match this comment style.

More documentation about the Ansible Configuration Settings file is
available in the official documentation:
https://docs.ansible.com/ansible/latest/reference_appendices/config.html

Signed-off-by: Nico Rikken <nico.rikken@alliander.com>