Repair tests
Setting check=True is a good sanity check. Previously some of these
commands were silently failing.

The thing that was really broken is related to the CVE-2022-39253 in
upstream Git. New versions of Git no longer allow `git submodule add
repository path` where repository is a file. We add a flag to explicitly
allow this.

See this URL for a meta analysis of the problem:

https://vielmetti.typepad.com/logbook/2022/10/git-security-fixes-lead-to-fatal-transport-file-not-allowed-error-in-ci-systems-cve-2022-39253.html

Signed-off-by: Carmen Bianca Bakker <carmenbianca@fsfe.org>