Treat untracked files in Git repos as "ignored"
When the Git VCS strategy is used, ignore files that are not tracked by
Git. Untracked files will not be published by commands like `git push`
or `git archive`, but can sometimes occur during development cycles. For
example, one could write a small script to explore possible new options,
without the intention of ever committing this script. Forcing a
developer in this case to either add license headers anyways or to add
these temporary files to a project-wide `.gitignore` file would likely
be seen as a hassle rather than a helpful hint.

As soon as a file is staged for committing in Git, it becomes tracked,
and therefore is checked again.