Allow year ranges to contain space around the dash
As shown in the examples of the SPDX specification
(https://spdx.github.io/spdx-spec/v2.3/file-information/#8.8), the a
copyright year range is not required to contain spaces around the dash
character.

This patch allows zero or more spaces on each side of the dash, making
all of the following year ranges being accepted by reuse:
"2023"
"2022-2023"
"2022 - 2023"
"2022-  2023"
"2022  -2023"

Reuse still writes the copyright in a normalized "yyyy - yyyy" format.

Closes #647