add additional license classifiers
This only concerns the metadata information for the Python package, picked up
by e.g. PyPI. Currently, the valid SPDX license expression we provide in the
license field isn't properly understood and displayed as a "proprietary or
other" license. We hope that with these classifiers, this display will be
closer to the truth.