Fix ignored path detection in ignored directories
Previously a sub-path of an ignored directory was not considered to be
ignored, e.g. is_ignored returned False for a file build/hello.py even
if _all_ignored_files contained Path("build").  Instead of just checking
if the path is in the ignored files we also have to check if it is
inside of an ignored directory.

Additionally this commit fixes a bug that surfaced: when building the
_all_ignored_files set all_files contains an empty string as its last
element. This leads to Path(".") being an element in _all_ignored_files,
i.e. ignore everything. Using all but the last element from all_files
fixes that.