Better error handling for DEP5 parse errors
Before, the Debian Copyright object was created lazily. When it is first
needed, it is created. The Projects that were passed to other processes
did not yet have Copyright objects (because they weren't yet needed, and
can't be pickled). However, this sometimes resulted in every process
separately trying to parse the DEP5 file, and failing.

Now, the DEP5 file is parsed immediately instead of being lazily
created. We circumvent the pickling problem by unsetting the attribute
and re-parsing the file in the process. This is a bit stupid, but it
shouldn't result in errors if the DEP5 file is not changed in the
meantime.

Signed-off-by: Carmen Bianca BAKKER <carmenbianca@fsfe.org>