Fix the link to spdx/tools-python in _license.py
There was a breaking change v0.7 -> v0.8 in spdx/tools-python, causing
the link to the original location to be broken.

Replace the master branch with the actual commit at the time when the
file was copied.

The chosen commit is 462e986, based on Github history (2017-06-15) and
the commit date of the _license.py file (2019-04-14).