Add support for Sapling as a VCS as well.
I introduced support to ignore the `.sl` directory already, but that did
not include support for the VCS interface.

As a VCS, Sapling uses effectively the same interface as Mercurial,
although notably it does not support Mercurial's `.hgignore` file, nor
does it invent a new one: it uses `.gitignore`.

I opted for copying the Mercurial implementation though, it looks
more adequate.