Add `lint --only <files...>` option
This option restricts the generated report to only the files listed on
the command line. It is useful for example in the pre-commit hook, in
order to not fail the entire commit if there are some issues in files
that aren't supposed to be committed yet anyways.

This option is implemented by checking all files that we know about as
usual, but only reporting on (and failing for) issues in files listed.

We have to check all files, otherwise we wouldn't be able to catch
certain types of errors reliably or at all. Specifically, the "unused
licenses" check would often raise false-positive errors: If we only look
at the *to-be-committed* files, and those don't use a particular
license, we would fail the check even if another file uses it (which was
already committed before, but not touched during the new commit).

This does however mean that untracked files could cause false-negatives,
but this was already the case before, and will (still) be caught by CI.

One alternative to this approach is to not check for unused licenses at
all if doing such a partial check, but this would mean that downstream
CI scripts would have to be adjusted to explicitly run a full `reuse
lint` instead of running the pre-commit hook over all repo files.
However, this would only lead to a lot of complications without any
meaningful benefit.