feat: lines_output based on ProjectReport
Update to the format_lines variant, now based on the ProjectReport to include
errors that relate to the licenses.

Add a property to the Project that resembles the LICENSES/ directory to prevent
duplication of resolving that.

No additional ordering is done as it would only slow things down and it can
easily be done by the user.

Signed-off-by: Nico Rikken <nico@nicorikken.eu>