Generate project object on demand
This is a lot cleaner, and has better performance benefits. Commands
like 'download' do not always require a project object. Now, if the code
never touches 'obj.project', the object is never created.

Apart from performance benefits, it also means that irrelevant errors
spawned from creation of the project object are sidestepped, because the
object is never instantiated (in some cases).

Signed-off-by: Carmen Bianca BAKKER <carmenbianca@fsfe.org>