Skip newline before python docstring
Reuse is not meant to be a python formatting tool, however when reuse
adds a header, it forcibly adds a space before the docstring. This
behavior does not align with dedicated python formatting tool
docformatter. Users should be able to format their code as they wish and
only use reuse to manage licenses.

When the style is set to Python and the first line in the file is a
docstring, do not add a space between the license and the docstring.
This enables uses to choose to have a space or not between the license
and the docstring.

Signed-off-by: Charlie Jenkins <charlie@rivosinc.com>