Move license below h1-header in Markdown
Linters like PyMarkdown complain about the license comment starting Markdown documents.  While the `reuse` tool still puts the license comment in line 1, we can showcase in this repository how to handle this situation through manual editing to both please the linter and stay REUSE-compliant.

Related to the discussion in issue #868.