Remove contact from CopyrightNotice, recognise year ranges anywhere
I removed the concept of 'contact' from CopyrightNotice because it was
not actually used anywhere. It could be re-added later as a computed
property that is equal to 'the text in *name* which appears between
angle brackets at the end'.

I also re-jigged the logic of grabbing year ranges in copyright
notices. Previously, years could only be at the start of a notice. Now,
they are grabbed anywhere.