Relicense workflow files to CC0-1.0
We want people to be able to reuse or draw inspiration from these files
to the fullest extent. On the REUSE website, the sample snippets are
already CC0-1.0.

Where there are non-FSFE authors for these files, I have determined that
their contributions have either disappeared (removed or substantively
morphed into something else), or do not meet the threshold of
copyrightability.